% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{cut_GO}
\alias{cut_GO}
\title{Prune the GO Graph Beginning from the Roots}
\usage{
cut_GO(GOids, level = 2, startingnodes = c("GO:0008150", "GO:0003674",
  "GO:0005575"))
}
\arguments{
\item{GOids}{character vector of GOIDs}

\item{level}{up to which level the terms should be deleted ? Defaults to 2.}

\item{startingnodes}{set the starting nodes. The specified level is relative 
to the starting nodes. The default is set to the ID representing the terms:
'BP', 'MF', 'CC'.}
}
\value{
The output is a truncated character vector of the entered GOIDs. 
The more general terms are removed.
}
\description{
The \code{cut_GO()} function takes a character vector of GOIDS as
input and removes the terms to a specified level in the DAG.
}
