% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{merge_n_clean}
\alias{merge_n_clean}
\title{Merge the dataframes of different stages and cut out general GO 
terms}
\usage{
merge_n_clean(diff_stage_tables, cut_level = 2, pval = 0.01,
  sort_by = "pvalSum", modus = "GO", show_enrich_only = TRUE)
}
\arguments{
\item{diff_stage_tables}{put in dataframes obtained by \code{term_enrichment} for several
stages as list}

\item{cut_level}{up to which level starting from the root the terms should 
be deleted ? Defaults to 2.}

\item{sort_by}{how should the terms be sorted for the ggplot script ? The
default is "pvalSum" and it sorts the terms according to the sum of p-values
over the different stages. The second option is "variance" which sorts the 
terms according to the biggest spread in p-values over the stages.}

\item{modus}{"GO", "PANTHER" or "KEGG"}

\item{show_enrich_only}{If set to TRUE, TERMs that are depleted over all
stages are ingnored for the output. Defaults to TRUE.}
}
\value{
The output is a list of two items.\cr
The first item identified by "table_all_stages_sign_IDs" is a dataframe which
contains all Ontology IDs that are significantly enriched or depleted in any
of the experimental stages. This dataframe is the input for the ggplot script
\cr
The second item identified by "order_by_pval" specifies the order of terms 
for the visualization in the ggplot script.
}
\description{
\code{merge_n_clean()} takes the the data output of 
\code{term_enrichment()} for different stages, merges them, cuts out more
general GO terms and subsets the merged tables by the GO terms that are 
significantly enriched or depleted in any of the entered stages.
}
