% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_analysis.R
\name{term_enrichment}
\alias{term_enrichment}
\title{Testing for Enrichment and Depletion of Ontology Terms in a Geneset}
\usage{
term_enrichment(parameters_list_exp, parameters_list_ref, stage, modus = "GO",
  twosidetest = "midpval", multCorrect = "bonferroni")
}
\arguments{
\item{parameters_list_exp}{the list that is created by 
\code{term_2_gene_count()} for your \emph{gene set of interest} e.g. 
down-regulated genes}

\item{parameters_list_ref}{the list that is created by 
\code{term_2_gene_count()} for your \emph{reference gene set}.}

\item{stage}{specify name of experiment.}

\item{modus}{"GO", "KEGG" or "PANTHER" analysis. Defaults to "GO".}

\item{twosidetest}{"midpval" or "simpledouble" according to PMID: 17182697.
Default is "midpval".}

\item{multCorrect}{"holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
"fdr", "none". Default is "bonferroni".}
}
\value{
\code{term_enrichment} creates a dataframe with information of 
the statistical analysis.\cr
The column named "k" represents the number of genes
in the experimental set with an annotation to the row ID.\cr 
The column named "m" represents the number 
of genes in the reference set with an annotation to the row ID.\cr
The columns "kcorrected", "whichmin", "pvl_R" and "pval_L" are essential
for the twosided test. Please see PMID: 17182697 and source code.\cr
The column "k_to_n" ratio gives the proportion of genes in the experimental
set that have an annotation to the ID specified in the row.\cr
The column "twosided.logpval" is the negative log10 of the twosided pvalue.
\cr
"E" stands for enriched and "D" for depleted.
}
\description{
\code{term_2_gene_count}  tests for enrichment or depletion of
ontology term as compared to randomly sample a set of genes in the reference
set of \code{length(your_favourite_experimental_set)}.
}
\details{
Please be careful when you set the background or reference set.
The background has to be set in a reasonable fashion to get an appropriate
answer to a predefined biological question.
}
